Storyboard changes and json tweaks
I tweaked the alignment on some cells so the look write when there is
only a one line description. I'm modifying the maakejson.go directly,
but makde a lot of tweaks to the one already built so I could see how
descriptions fit and wrapped before saving them to the makejson.go file.