iOS7 Tab Bar icons / Map size
Set the map size and position on the Nearby screen to sit under the
Navigation Bar, now that transparency is involved.

I'm updating the Tab Bar icons - styled for iOS7 with empty and filled
versions - but there currently doesn't seem to be a way to set the Tab
Bar Select image through the XCode UI in a nice, easy, clicky way.
Still need to do subway.